pyproject.toml
[tool.poetry]
name = "ecoscan"
version = "0.1.0"
description = ""
authors = ["Sarthak5598 <sarthak5598sharma@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
streamlit = "^1.40.1"
opencv-python = "^4.10.0.84"
av = "^13.1.0"
streamlit-camera-input-live = "^0.2.0"
openai = "^1.54.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
